backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Faker==18.3.1
pandas==2.0.0
numpy==1.24.2
matplotlib==3.7.1
plotly==5.14.1
sqlalchemy==2.0.9
transformers==4.27.4
openai==0.27.4
python-dotenv==1.0.0